pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API repository."
authors = [ "Niels Degrande <niels@degran.de>" ]
license = "Other/Proprietary"
readme = "README.md"
packages = []

[tool.poetry.scripts]
api = "api.__main__:main"

[tool.poetry.dependencies]
python = "~3.11"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
bcrypt = "^4.1.2"
coloredlogs = "~15.0"
fastapi = "^0.105.0"
google-cloud-storage = "^2.14.0"
python-box = "~7.1"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.23"
uvicorn = "^0.24.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
httpx = "^0.25.2"
pre-commit = "~3.6"
pytest = "~7.4"
pytest-asyncio = "^0.23.2"
pytest-cov = "~4.1"
pytest-mock = "~3.12"
pytest-timeout = "~2.2"
pytest-xdist = "~3.5"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
add-trailing-comma = "*"
black = "*"
codespell = "*"
ipython = "*"
isort = "*"
pre-commit = "*"
pyright = "*"
ruff = "*"
sqlalchemy2-stubs = "*"
sqlfluff = "*"
types-PyYAML = "*"
vulture = "*"

[tool.pytest.ini_options]
timeout = "60"
addopts = [
    "--cov-report=html:test-reports/htmlcov",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=xml:test-reports/coverage.xml",
    "--cov=api",
    "--junitxml=test-reports/junit/junit.xml",
    "--no-cov-on-fail",
    "--strict-markers",
    "--verbose",
    "-r sxX",
]
cache_dir = ".cache"
testpaths = "tests"

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_also = [
    # Do not complain if tests do not hit defensive assertion code.
    "raise AssertionError",
    "raise NotImplementedError",

    # Do not complain if non-runnable code is not run.
    "if __name__ == .__main__.:",
]
ignore_errors = true

[tool.coverage.html]
directory = "coverage_html_report"

[tool.ruff]
target-version = "py311"
select = [ "ALL" ]
ignore = [
    "D203", # one-blank-line-before-class (incompatible with D211).
    "D213", # multi-line-summary-second-line (incompatible with D212).
]

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"

[tool.ruff.per-file-ignores]
# Do not complain if __init__.py files have unused imports.
"__init__.py" = [ "F401" ]
# Do not complain if tests use the assert keyword.
"tests/*" = [ "S101" ]
